requirements.txt
flask
flask-bootstrap
flask-wtf
flask-login
firebase-admin
python-dotenv
